修復浮點數序列化精度問題
- 修改JSON序列化設置，控制浮點數的顯示格式
- 統一使用float類型表示cfg_rescale屬性
- 明確指定浮點數字面量類型為float (0f)
- 確保所有序列化方法使用相同的設置

這些修改解決了CFG值為0.3時，實際顯示為0.300000000000000很長一串數字的問題。

+++ body/NovalAIBase.cs
@@ -128,7 +128,7 @@ namespace AutoNai3Tools.body {
         public double controlnet_strength { get; set; }
         public bool legacy { get; set; }
         public bool add_original_image { get; set; }
-        public double cfg_rescale { get; set; }
+        public float cfg_rescale { get; set; }
         public string noise_schedule { get; set; }
         public string image { get; set; }
         public float? strength { get; set; }
